Fix: move a visitor test to a visitor test file
