+++ packages/appsync-modelgen-plugin/src/__tests__/utils/process-connections.test.ts
@@ -1,51 +1,13 @@
 import {
   processConnections,
   CodeGenConnectionType,
-  CodeGenFieldConnection,
   CodeGenFieldConnectionHasMany,
   CodeGenFieldConnectionBelongsTo,
   CodeGenFieldConnectionHasOne,
   getConnectedField,
 } from '../../utils/process-connections';
-import { CodeGenModelMap, CodeGenGenerateEnum, AppSyncModelVisitor } from '../../visitors/appsync-visitor';
+import { CodeGenModelMap } from '../../visitors/appsync-visitor';
 import { processConnectionsV2 } from '../../utils/process-connections-v2';
-import { buildSchema, parse, visit } from 'graphql';
-import { directives, scalars } from '../../scalars/supported-directives';
-import { SWIFT_SCALAR_MAP } from '../../scalars';
-
-const buildSchemaWithDirectives = (schema: String) => {
-  return buildSchema([schema, directives, scalars].join('\n'));
-};
-
-const getPipelinedTransformerVisitor = (
-  schema: string,
-  selectedType?: string,
-  generate: CodeGenGenerateEnum = CodeGenGenerateEnum.code,
-  isTimestampFieldsAdded: boolean = true,
-  emitAuthProvider: boolean = true,
-  generateIndexRules: boolean = true,
-  handleListNullabilityTransparently: boolean = true,
-  usePipelinedTransformer: boolean = true
-) => {
-  const ast = parse(schema);
-  const builtSchema = buildSchemaWithDirectives(schema);
-  const visitor = new AppSyncModelVisitor(
-    builtSchema,
-    {
-      directives,
-      target: 'swift',
-      scalars: SWIFT_SCALAR_MAP,
-      isTimestampFieldsAdded,
-      emitAuthProvider,
-      generateIndexRules,
-      handleListNullabilityTransparently,
-      usePipelinedTransformer: usePipelinedTransformer
-    },
-    { selectedType, generate },
-  );
-  visit(ast, { leave: visitor });
-  return visitor;
-};
 
 describe('process connection', () => {
   describe('Bi-Directional connection (named connection)', () => {
@@ -810,179 +772,5 @@ describe('process connection', () => {
       });
     });
 
-    describe('manyToMany testing', () => {
-      let simpleManyToManySchema;
-      let simpleManyModelMap;
-      let transformedSimpleManyModelMap;
-
-      beforeEach(() => {
-        simpleManyToManySchema = /* GraphQL */ `
-          type Human @model {
-            governmentID: ID! @primaryKey
-            pets: [Animal] @manyToMany(relationName: "PetFriend")
-          }
-          
-          type Animal @model {
-            animalTag: ID!
-            humanFriend: [Human] @manyToMany(relationName: "PetFriend")
-          }
-        `;
-
-        simpleManyModelMap = {
-          Human: {
-            name: 'Human',
-            type: 'model',
-            directives: [],
-            fields: [
-              {
-                type: 'ID',
-                isNullable: false,
-                isList: false,
-                name: 'governmentID',
-                directives: [{ name: 'primaryKey', arguments: {} }]
-              },
-              {
-                type: 'Animal',
-                isNullable: true,
-                isList: true,
-                name: 'pets',
-                directives: [{ name: 'manyToMany', arguments: { relationName: 'PetFriend' } }]
-              }
-            ]
-          },
-          Animal: {
-            name: 'Animal',
-            type: 'model',
-            directives: [],
-            fields: [
-              {
-                type: 'ID',
-                isNullable: false,
-                isList: false,
-                name: 'animalTag',
-                directives: [],
-              },
-              {
-                type: 'Human',
-                isNullable: true,
-                isList: true,
-                name: 'postID',
-                directives: [{ name: 'manyToMany', arguments: { relationName: 'PetFriend' } }]
-              }
-            ],
-          }
-        };
-      });
-
-      transformedSimpleManyModelMap = {
-        Human: {
-          name: 'Human',
-          type: 'model',
-          directives: [{name: 'model', arguments: {}}],
-          fields: [
-            {
-              type: 'ID',
-              isNullable: false,
-              isList: false,
-              name: 'governmentID',
-              directives: [{ name: 'primaryKey', arguments: {} }]
-            },
-            {
-              type: 'PetFriend',
-              isNullable: true,
-              isList: true,
-              name: 'pets',
-              directives: [{ name: 'hasMany', arguments: { fields: ['governmentID'] } }]
-            }
-          ]
-        },
-        Animal: {
-          name: 'Animal',
-          type: 'model',
-          directives: [{name: 'model', arguments: {}}],
-          fields: [
-            {
-              type: 'ID',
-              isNullable: false,
-              isList: false,
-              name: 'animalTag',
-              directives: [],
-            },
-            {
-              type: 'PetFriend',
-              isNullable: true,
-              isList: true,
-              name: 'postID',
-              directives: [{ name: 'hasMany', arguments: { fields: ['id'] } }]
-            }
-          ],
-        },
-        PetFriend: {
-          name: 'PetFriend',
-          type: 'model',
-          directives: [{name: 'model', arguments: {}}],
-          fields: [
-            {
-              type: 'ID',
-              isNullable: false,
-              isList: false,
-              name: 'id',
-              directives: []
-            },
-            {
-              type: 'ID',
-              isNullable: false,
-              isList: false,
-              name: 'humanID',
-              directives: [{ name: 'index', arguments: { name: 'byHuman', sortKeyFields: ['animalID'] } }]
-            },
-            {
-              type: 'ID',
-              isNullable: false,
-              isList: false,
-              name: 'animalID',
-              directives: [{ name: 'index', arguments: { name: 'byAnimal', sortKeyFields: ['humanID'] } }]
-            },
-            {
-              type: 'Human',
-              isNullable: false,
-              isList: false,
-              name: 'human',
-              directives: [{ name: 'belongsTo', arguments: { fields: ['humanID'] } }]
-            },
-            {
-              type: 'Animal',
-              isNullable: false,
-              isList: false,
-              name: 'animal',
-              directives: [{ name: 'belongsTo', arguments: { fields: ['humanID'] } }]
-            }
-          ]
-        }
-      };
-
-      it('Should correctly convert the model map of a simple manyToMany', () => {
-        const visitor = getPipelinedTransformerVisitor(simpleManyToManySchema);
-        let code = visitor.generate();
-
-        expect(visitor.models.Human.fields.length).toEqual(5);
-        expect(visitor.models.Human.fields[2].directives[0].name).toEqual('hasMany')
-        expect(visitor.models.Human.fields[2].directives[0].arguments.fields.length).toEqual(1)
-        expect(visitor.models.Human.fields[2].directives[0].arguments.fields[0]).toEqual('governmentID')
-        expect(visitor.models.Human.fields[2].directives[0].arguments.indexName).toEqual('byHuman')
-        expect(visitor.models.PetFriend).toBeDefined();
-        expect(visitor.models.PetFriend.fields.length).toEqual(5);
-        expect(visitor.models.PetFriend.fields[2].directives[0].name).toEqual('belongsTo')
-        expect(visitor.models.PetFriend.fields[2].directives[0].arguments.fields.length).toEqual(1)
-        expect(visitor.models.PetFriend.fields[2].directives[0].arguments.fields[0]).toEqual('animalID')
-        expect(visitor.models.Animal.fields.length).toEqual(5);
-        expect(visitor.models.Animal.fields[2].type).toEqual('PetFriend');
-        expect(visitor.models.Animal.fields[2].directives.length).toEqual(1);
-        expect(visitor.models.Animal.fields[2].directives[0].name).toEqual('hasMany');
-        expect(visitor.models.Animal.fields[2].directives[0].arguments.fields.length).toEqual(1)
-        expect(visitor.models.Animal.fields[2].directives[0].arguments.fields[0]).toEqual('id')
-        expect(visitor.models.Animal.fields[2].directives[0].arguments.indexName).toEqual('byAnimal')
-      });
-    });
   });
 });
